mcp/search_mcp/requirements.txt
clarifai==11.4.7
anyio==4.9.0
mcp==1.9.0
fastmcp==2.3.4
requests>=2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-search-results==2.4.2 
newspaper3k